ODLC_src/alphanumeric_classifier/requirements.txt
requests 
numpy 
torch 
pillow 
easyocr 
opencv-python 
matplotlib